requirements.txt
python-dotenv
langchain
openai==1.0.0
